modifications 1. featuerDetection: centroid mode handling 2. featureAlignment: finding matched features is modified to find most plausible one 3. featureToMs2: moving mzXML files is modified

+++ featureToMS2.py
@@ -230,6 +230,9 @@ def ms2ForFeatures(full, mzxmlFiles, paramFile):
     for file in mzxmlFiles:
         baseFilename = os.path.basename(file)
         featureDirectory = os.path.join(os.getcwd(), os.path.splitext(baseFilename)[0])
-        os.rename(file, os.path.join(featureDirectory, baseFilename))
+        if os.path.isfile(os.path.join(featureDirectory, baseFilename)):
+            continue
+        else:
+            os.rename(file, os.path.join(featureDirectory, baseFilename))
 
     return df, featureToScan
